Move main console code to TMainConsole.h / TMainConsole.cpp
