+++ src/ActionUnit.cpp
@@ -25,7 +25,6 @@
 
 #include "TAction.h"
 #include "TCommandLine.h"
-#include "TConsole.h"
 #include "TEasyButtonBar.h"
 #include "TToolBar.h"
 #include "mudlet.h"
