+++ src/TBuffer.h
@@ -312,7 +312,7 @@ inline QDebug& operator<<(QDebug& debug, const TChar::AttributeFlags& attributes
     if (attributes & TChar::Echo) {
         presentAttributes << QLatin1String("Echo (0x100)");
     }
-    result.append(presentAttributes.join(", "));
+    result.append(presentAttributes.join(QLatin1String(", ")));
     result.append(QLatin1String(")"));
     debug.nospace() << result;
     return debug;
